requirements.txt
aiohttp>=3.9
websockets>=12.0
asyncpg>=0.29
python-dotenv>=1.0
